Clean up typings for Direction / SplitDirectoin
