+++ browser/src/Services/Sidebar/index.ts
@@ -18,7 +18,7 @@ export const activate = (configuration: Configuration, workspace: Workspace) => 
     _sidebarManager = new SidebarManager()
 
     if (configuration.getValue("experimental.sidebar.enabled")) {
-        const leftDock = windowManager.getDock(2)
+        const leftDock = windowManager.getDock("left")
         leftDock.addSplit(new SidebarSplit(_sidebarManager))
         leftDock.addSplit(new SidebarContentSplit(_sidebarManager))
